refactor(root): rename encryption files to avoid mistype ss-17
