+++ apps/backend/src/libs/modules/encryption/libs/base-encryption.module.ts
