+++ apps/backend/src/libs/modules/encryption/libs/types/encryption-data-response.type.ts
